Move 'determine_qc_protocol' function from qc/illuminaqc.py to qc/utils.py.
